requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
jinja2
python-multipart
passlib[bcrypt]
python-dotenv
